added database functionality to multer
