+++ api/controllers/uploads.js
@@ -1,42 +1,24 @@
 // eslint-disable-next-line no-unused-vars
-const uploadSchema = require('../models/uploads');
-const multer = require('multer');
-const path = require('path');
-
-const storage = multer.diskStorage({
-  destination: '../uploads',
-  filename: (req, file, cb) => {
-    cb(
-      null,
-      file.fieldname + '-' + Date.now() + path.extname(file.originalname)
-    );
-  },
-});
-
-function fileFilter(req, file, cb) {
-  const fileTypes = '/jpeg|jpg|png/';
-  const extname = fileTypes.test(path.extname(file.originalname).toLowerCase());
-  const mimetype = fileTypes.test(file.mimetype);
-  if (mimetype && extname) {
-    return cb(null, true);
-  } else {
-    cb('Error wrong format');
-  }
-}
-const upload = multer({
-  storage: storage,
-  fileFilter: (req, file, cb) => {
-    fileFilter(file, cb);
-  },
-});
-
-exports.uploaded = (req, res, next) => {
-  upload.single('picture');
-  next();
-};
-
-// eslint-disable-next-line no-unused-vars
-exports.storedPicture = (req, res, next) => {
+const Upload = require('../models/uploads');
+function storedPicture(req, res) {
   //add database logic
-  console.log('hii');
-};
+
+  const image = new Upload({
+    name: req.body.name,
+    picture: req.file.path,
+  });
+  image
+    .save()
+    .then((results) => {
+      //sending appropriate status
+      console.log(results);
+      res.status(201).json({
+        message: 'Handling POST requests to /uploads/',
+      });
+    })
+    .catch((err) => {
+      console.log(err);
+      res.status(500).json({ error: err });
+    });
+}
+module.exports = storedPicture;
